Support TLS 1.3
Enables support for negotiating TLS 1.3.  TLS 1.3 is not enabled
unless SSLContext.TLSv1.3 is requested or setEnabledProtocols() is
called with a set of values that includes TLSv1.3.

Detailed changes:

Adds protocol constants for TLS 1.3, and provides SSLContext.TLSv1.3,
which has TLS 1.3 enabled by default.

Adjusts cipher suite code for TLS 1.3 suites.  When enabled, all TLS
1.3 cipher suites are always returned from supportedCipherSuites() and
enabledCipherSuites().  Attempts to customize TLS 1.3 cipher suites in
setEnabledCipherSuites() are ignored.

Splits {SSLEngine,SSLSocket}Test into version-dependent and
version-independent tests.  The latter remain in
{SSLEngine,SSLSocket}Test and the former move into new files
{SSLEngine,SSLSocket}VersionCompatibilityTest, which are parameterized
to test all combinations of client and server on TLS 1.2 and TLS 1.3.

Manufactures session IDs from tickets for sessions which have no ID
but do have a ticket (which in practice is TLS 1.3 sessions), so that
we can cache them properly.

Remove a pile of RI-specific TLS-related expectation declarations from
StandardNames.  We don't actually verify the behavior of the RI at any
point, so it was just making the code more confusing.

Fixes #479.
