+++ CAPABILITIES.md
@@ -9,57 +9,73 @@ and other identifiers that are supported by Conscrypt.
 
 ### Protocol Versions
 
-* SSLv3 (ignored)
-* TLSv1
-* TLSv1.1
-* TLSv1.2
+* `SSLv3` (ignored)
+* `TLSv1`
+* `TLSv1.1`
+* `TLSv1.2`
+* `TLSv1.3`
 
-Conscrypt supports TLS v1.0-1.2.  For backwards compatibility it will accept "SSLv3"
-in calls to methods like
-[setEnabledProtocols()](https://docs.oracle.com/javase/9/docs/api/javax/net/ssl/SSLSocket.html#setEnabledProtocols-java.lang.String:A-)
-but will ignore it.
+Conscrypt supports TLS v1.0-1.3.  For backwards compatibility it will accept
+`SSLv3` in calls to methods like
+[`setEnabledProtocols()`](https://docs.oracle.com/javase/9/docs/api/javax/net/ssl/SSLSocket.html#setEnabledProtocols-java.lang.String:A-)
+but will ignore it.  TLS 1.3 is not enabled unless `SSLContext.TLSv1.3` is
+requested or it's explicitly enabled by a call to
+[`setEnabledProtocols()`](https://docs.oracle.com/javase/9/docs/api/javax/net/ssl/SSLSocket.html#setEnabledProtocols-java.lang.String:A-).
 
 ### SSLContext
 
-* Default
-* SSL
-* TLS
-* TLSv1
-* TLSv1.1
-* TLSv1.2
+* `Default`
+* `SSL`
+* `TLS`
+* `TLSv1`
+* `TLSv1.1`
+* `TLSv1.2`
+* `TLSv1.3`
 
-Conscrypt provides the above set of SSLContext algorithm names for JSSE purposes, including
-the special value "Default", which is used to determine the value of
-[SSLContext.getDefault()](https://docs.oracle.com/javase/9/docs/api/javax/net/ssl/SSLContext.html#getDefault--).
-All of these values return a context where TLS v1.0-1.2 are all enabled.
+Conscrypt provides the above set of SSLContext algorithm names for JSSE
+purposes, including the special value `Default`, which is used to determine the
+value of
+[`SSLContext.getDefault()`](https://docs.oracle.com/javase/9/docs/api/javax/net/ssl/SSLContext.html#getDefault--).
+All of these values return a context where TLS v1.0-1.2 are all enabled except
+`TLSv1.3`, which has TLS v1.0-1.3 enabled.
 
 ### Cipher Suites
 
 #### Enabled
-* TLS_ECDHE_ECDSA_WITH_AES_128_CBC_SHA
-* TLS_ECDHE_ECDSA_WITH_AES_128_GCM_SHA256
-* TLS_ECDHE_ECDSA_WITH_AES_256_CBC_SHA
-* TLS_ECDHE_ECDSA_WITH_AES_256_GCM_SHA384
-* TLS_ECDHE_ECDSA_WITH_CHACHA20_POLY1305_SHA256
-* TLS_ECDHE_RSA_WITH_AES_128_CBC_SHA
-* TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256
-* TLS_ECDHE_RSA_WITH_AES_256_CBC_SHA
-* TLS_ECDHE_RSA_WITH_AES_256_GCM_SHA384
-* TLS_ECDHE_RSA_WITH_CHACHA20_POLY1305_SHA256
-* TLS_RSA_WITH_AES_128_CBC_SHA
-* TLS_RSA_WITH_AES_128_GCM_SHA256
-* TLS_RSA_WITH_AES_256_CBC_SHA
-* TLS_RSA_WITH_AES_256_GCM_SHA384
+* TLS 1.2
+  * `TLS_ECDHE_ECDSA_WITH_AES_128_CBC_SHA`
+  * `TLS_ECDHE_ECDSA_WITH_AES_128_GCM_SHA256`
+  * `TLS_ECDHE_ECDSA_WITH_AES_256_CBC_SHA`
+  * `TLS_ECDHE_ECDSA_WITH_AES_256_GCM_SHA384`
+  * `TLS_ECDHE_ECDSA_WITH_CHACHA20_POLY1305_SHA256`
+  * `TLS_ECDHE_RSA_WITH_AES_128_CBC_SHA`
+  * `TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256`
+  * `TLS_ECDHE_RSA_WITH_AES_256_CBC_SHA`
+  * `TLS_ECDHE_RSA_WITH_AES_256_GCM_SHA384`
+  * `TLS_ECDHE_RSA_WITH_CHACHA20_POLY1305_SHA256`
+  * `TLS_RSA_WITH_AES_128_CBC_SHA`
+  * `TLS_RSA_WITH_AES_128_GCM_SHA256`
+  * `TLS_RSA_WITH_AES_256_CBC_SHA`
+  * `TLS_RSA_WITH_AES_256_GCM_SHA384`
+* TLS 1.3
+  * `TLS_AES_128_GCM_SHA256`
+  * `TLS_AES_256_GCM_SHA384`
+  * `TLS_CHACHA20_POLY1305_SHA256`
 
-The above cipher suites are enabled by default.
+The above cipher suites are enabled by default when the associated version of
+the protocol is enabled.  The TLS 1.3 cipher suites cannot be customized; they
+are always enabled when TLS 1.3 is enabled, and any attempt to disable them via
+a call to
+[`setEnabledCipherSuites()`](https://docs.oracle.com/javase/9/docs/api/javax/net/ssl/SSLSocket.html#setEnabledCipherSuites-java.lang.String:A-)
+is ignored.
 
 #### Supported But Not Enabled
-* SSL_RSA_WITH_3DES_EDE_CBC_SHA
-* TLS_ECDHE_PSK_WITH_AES_128_CBC_SHA
-* TLS_ECDHE_PSK_WITH_AES_256_CBC_SHA
-* TLS_ECDHE_PSK_WITH_CHACHA20_POLY1305_SHA256
-* TLS_PSK_WITH_AES_128_CBC_SHA
-* TLS_PSK_WITH_AES_256_CBC_SHA
+* `SSL_RSA_WITH_3DES_EDE_CBC_SHA`
+* `TLS_ECDHE_PSK_WITH_AES_128_CBC_SHA`
+* `TLS_ECDHE_PSK_WITH_AES_256_CBC_SHA`
+* `TLS_ECDHE_PSK_WITH_CHACHA20_POLY1305_SHA256`
+* `TLS_PSK_WITH_AES_128_CBC_SHA`
+* `TLS_PSK_WITH_AES_256_CBC_SHA`
 
 The above cipher suites are supported, but not enabled by default.
