+++ common/src/jni/main/cpp/conscrypt/native_crypto.cc
@@ -8577,6 +8577,37 @@ static void NativeCrypto_SSL_free(JNIEnv* env, jclass, jlong ssl_address, CONSCR
     SSL_free(ssl);
 }
 
+static jbyteArray get_session_id(JNIEnv* env, SSL_SESSION* ssl_session) {
+    jbyteArray result = nullptr;
+    unsigned int length;
+    const uint8_t* id = SSL_SESSION_get_id(ssl_session, &length);
+    JNI_TRACE("ssl_session=%p get_session_id id=%p length=%u", ssl_session, id, length);
+    if (id && length > 0) {
+        result = env->NewByteArray(static_cast<jsize>(length));
+        if (result != nullptr) {
+            const jbyte* src = reinterpret_cast<const jbyte*>(id);
+            env->SetByteArrayRegion(result, 0, static_cast<jsize>(length), src);
+        }
+    } else if (SSL_SESSION_has_ticket(ssl_session)) {
+        JNI_TRACE("ssl_session=%p get_session_id has_ticket=true", ssl_session);
+        const uint8_t* ticket;
+        size_t tick_len;
+        SSL_SESSION_get0_ticket(ssl_session, &ticket, &tick_len);
+        uint8_t data[32];
+        if (!EVP_Digest(ticket, tick_len, data, &length, EVP_sha256(), NULL)) {
+            return nullptr;
+        }
+        result = env->NewByteArray(static_cast<jsize>(length));
+        if (result != nullptr) {
+            const jbyte* src = reinterpret_cast<const jbyte*>(data);
+            env->SetByteArrayRegion(result, 0, static_cast<jsize>(length), src);
+        }
+    } else {
+        JNI_TRACE("ssl_session=%p get_session_id has_ticket=false", ssl_session);
+    }
+    return result;
+}
+
 /**
  * Gets and returns in a byte array the ID of the actual SSL session.
  */
@@ -8588,15 +8619,8 @@ static jbyteArray NativeCrypto_SSL_SESSION_session_id(JNIEnv* env, jclass,
     if (ssl_session == nullptr) {
         return nullptr;
     }
-    unsigned length;
-    const uint8_t* id = SSL_SESSION_get_id(ssl_session, &length);
-    jbyteArray result = env->NewByteArray(static_cast<jsize>(length));
-    if (result != nullptr) {
-        const jbyte* src = reinterpret_cast<const jbyte*>(id);
-        env->SetByteArrayRegion(result, 0, static_cast<jsize>(length), src);
-    }
-    JNI_TRACE("ssl_session=%p NativeCrypto_SSL_SESSION_session_id => %p length=%d", ssl_session,
-              result, length);
+    jbyteArray result = get_session_id(env, ssl_session);
+    JNI_TRACE("ssl_session=%p NativeCrypto_SSL_SESSION_session_id => %p", ssl_session, result);
     return result;
 }
 
@@ -8731,16 +8755,8 @@ static jbyteArray NativeCrypto_SSL_session_id(JNIEnv* env, jclass, jlong ssl_add
     if (ssl_session == nullptr) {
         return nullptr;
     }
-
-    unsigned session_id_length;
-    const uint8_t* session_id = SSL_SESSION_get_id(ssl_session, &session_id_length);
-    jbyteArray result = env->NewByteArray(static_cast<jsize>(session_id_length));
-    if (result != nullptr) {
-        const jbyte* src = reinterpret_cast<const jbyte*>(session_id);
-        env->SetByteArrayRegion(result, 0, static_cast<jsize>(session_id_length), src);
-    }
-    JNI_TRACE("ssl_session=%p NativeCrypto_SSL_session_id => %p session_id_length=%d", ssl_session,
-              result, session_id_length);
+    jbyteArray result = get_session_id(env, ssl_session);
+    JNI_TRACE("ssl_session=%p NativeCrypto_SSL_session_id => %p", ssl_session, result);
     return result;
 }
