+++ common/src/main/java/org/conscrypt/SSLParametersImpl.java
@@ -25,6 +25,7 @@ import java.security.SecureRandom;
 import java.security.UnrecoverableKeyException;
 import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.Set;
 import javax.crypto.SecretKey;
 import javax.net.ssl.KeyManager;
 import javax.net.ssl.KeyManagerFactory;
@@ -202,6 +203,10 @@ final class SSLParametersImpl implements Cloneable {
      * @return the names of enabled cipher suites
      */
     String[] getEnabledCipherSuites() {
+        if (Arrays.asList(enabledProtocols).contains(NativeCrypto.SUPPORTED_PROTOCOL_TLSV1_3)) {
+            return SSLUtils.concat(
+                    NativeCrypto.SUPPORTED_TLS_1_3_CIPHER_SUITES, enabledCipherSuites);
+        }
         return enabledCipherSuites.clone();
     }
 
@@ -209,7 +214,9 @@ final class SSLParametersImpl implements Cloneable {
      * Sets the enabled cipher suites after filtering through OpenSSL.
      */
     void setEnabledCipherSuites(String[] cipherSuites) {
-        enabledCipherSuites = NativeCrypto.checkEnabledCipherSuites(cipherSuites).clone();
+        enabledCipherSuites = NativeCrypto.checkEnabledCipherSuites(
+                filterFromCipherSuites(cipherSuites,
+                        NativeCrypto.SUPPORTED_TLS_1_3_CIPHER_SUITES_SET));
     }
 
     /**
@@ -381,6 +388,19 @@ final class SSLParametersImpl implements Cloneable {
             }
         }
         return newProtocols.toArray(EMPTY_STRING_ARRAY);
+    }
+
+    private static String[] filterFromCipherSuites(String[] cipherSuites, Set<String> toRemove) {
+        if (cipherSuites == null) {
+            return cipherSuites;
+        }
+        ArrayList<String> newCipherSuites = new ArrayList<String>();
+        for (String cipherSuite : cipherSuites) {
+            if (!toRemove.contains(cipherSuite)) {
+                newCipherSuites.add(cipherSuite);
+            }
+        }
+        return newCipherSuites.toArray(EMPTY_STRING_ARRAY);
     }
 
     private static final String[] EMPTY_STRING_ARRAY = new String[0];
@@ -581,39 +601,25 @@ final class SSLParametersImpl implements Cloneable {
                 // higher priority than X.509 cipher suites.
                 // NOTE: There are cipher suites that use both X.509 and PSK (e.g., those based on
                 // RSA_PSK key exchange). However, these cipher suites are not currently supported.
-                return concat(
+                return SSLUtils.concat(
                         NativeCrypto.DEFAULT_PSK_CIPHER_SUITES,
                         NativeCrypto.DEFAULT_X509_CIPHER_SUITES,
                         new String[] {NativeCrypto.TLS_EMPTY_RENEGOTIATION_INFO_SCSV});
             } else {
                 // Only X.509 cipher suites need to be listed.
-                return concat(
+                return SSLUtils.concat(
                         NativeCrypto.DEFAULT_X509_CIPHER_SUITES,
                         new String[] {NativeCrypto.TLS_EMPTY_RENEGOTIATION_INFO_SCSV});
             }
         } else if (pskCipherSuitesNeeded) {
             // Only PSK cipher suites need to be listed.
-            return concat(
+            return SSLUtils.concat(
                     NativeCrypto.DEFAULT_PSK_CIPHER_SUITES,
                     new String[] {NativeCrypto.TLS_EMPTY_RENEGOTIATION_INFO_SCSV});
         } else {
             // Neither X.509 nor PSK cipher suites need to be listed.
             return new String[] {NativeCrypto.TLS_EMPTY_RENEGOTIATION_INFO_SCSV};
         }
-    }
-
-    private static String[] concat(String[]... arrays) {
-        int resultLength = 0;
-        for (String[] array : arrays) {
-            resultLength += array.length;
-        }
-        String[] result = new String[resultLength];
-        int resultOffset = 0;
-        for (String[] array : arrays) {
-            System.arraycopy(array, 0, result, resultOffset, array.length);
-            resultOffset += array.length;
-        }
-        return result;
     }
 
     /**
