+++ constants/src/gen/cpp/generate_constants.cc
@@ -75,6 +75,7 @@ int main(int /* argc */, char ** /* argv */) {
   CONST(TLS1_VERSION);
   CONST(TLS1_1_VERSION);
   CONST(TLS1_2_VERSION);
+  CONST(TLS1_3_VERSION);
 
   CONST(SSL_SENT_SHUTDOWN);
   CONST(SSL_RECEIVED_SHUTDOWN);
