+++ openjdk-integ-tests/src/test/java/org/conscrypt/ConscryptJava6Suite.java
@@ -45,6 +45,7 @@ import org.conscrypt.javax.net.ssl.KeyStoreBuilderParametersTest;
 import org.conscrypt.javax.net.ssl.SNIHostNameTest;
 import org.conscrypt.javax.net.ssl.SSLContextTest;
 import org.conscrypt.javax.net.ssl.SSLEngineTest;
+import org.conscrypt.javax.net.ssl.SSLEngineVersionCompatibilityTest;
 import org.conscrypt.javax.net.ssl.SSLParametersTest;
 import org.conscrypt.javax.net.ssl.SSLServerSocketFactoryTest;
 import org.conscrypt.javax.net.ssl.SSLServerSocketTest;
@@ -52,6 +53,7 @@ import org.conscrypt.javax.net.ssl.SSLSessionContextTest;
 import org.conscrypt.javax.net.ssl.SSLSessionTest;
 import org.conscrypt.javax.net.ssl.SSLSocketFactoryTest;
 import org.conscrypt.javax.net.ssl.SSLSocketTest;
+import org.conscrypt.javax.net.ssl.SSLSocketVersionCompatibilityTest;
 import org.conscrypt.javax.net.ssl.TrustManagerFactoryTest;
 import org.conscrypt.javax.net.ssl.X509KeyManagerTest;
 import org.junit.BeforeClass;
@@ -94,6 +96,7 @@ import org.junit.runners.Suite;
         SNIHostNameTest.class,
         SSLContextTest.class,
         SSLEngineTest.class,
+        SSLEngineVersionCompatibilityTest.class,
         SSLParametersTest.class,
         SSLServerSocketFactoryTest.class,
         SSLServerSocketTest.class,
@@ -101,6 +104,7 @@ import org.junit.runners.Suite;
         SSLSessionTest.class,
         SSLSocketFactoryTest.class,
         SSLSocketTest.class,
+        SSLSocketVersionCompatibilityTest.class,
         TrustManagerFactoryTest.class,
         X509KeyManagerTest.class,
 })
