+++ openjdk-integ-tests/src/test/java/org/conscrypt/ConscryptSuite.java
@@ -48,6 +48,7 @@ import org.conscrypt.javax.net.ssl.KeyStoreBuilderParametersTest;
 import org.conscrypt.javax.net.ssl.SNIHostNameTest;
 import org.conscrypt.javax.net.ssl.SSLContextTest;
 import org.conscrypt.javax.net.ssl.SSLEngineTest;
+import org.conscrypt.javax.net.ssl.SSLEngineVersionCompatibilityTest;
 import org.conscrypt.javax.net.ssl.SSLParametersTest;
 import org.conscrypt.javax.net.ssl.SSLServerSocketFactoryTest;
 import org.conscrypt.javax.net.ssl.SSLServerSocketTest;
@@ -55,6 +56,7 @@ import org.conscrypt.javax.net.ssl.SSLSessionContextTest;
 import org.conscrypt.javax.net.ssl.SSLSessionTest;
 import org.conscrypt.javax.net.ssl.SSLSocketFactoryTest;
 import org.conscrypt.javax.net.ssl.SSLSocketTest;
+import org.conscrypt.javax.net.ssl.SSLSocketVersionCompatibilityTest;
 import org.conscrypt.javax.net.ssl.TrustManagerFactoryTest;
 import org.conscrypt.javax.net.ssl.X509KeyManagerTest;
 import org.junit.BeforeClass;
@@ -96,6 +98,7 @@ import org.junit.runners.Suite;
         SNIHostNameTest.class,
         SSLContextTest.class,
         SSLEngineTest.class,
+        SSLEngineVersionCompatibilityTest.class,
         SSLParametersTest.class,
         SSLServerSocketFactoryTest.class,
         SSLServerSocketTest.class,
@@ -103,6 +106,7 @@ import org.junit.runners.Suite;
         SSLSessionTest.class,
         SSLSocketFactoryTest.class,
         SSLSocketTest.class,
+        SSLSocketVersionCompatibilityTest.class,
         TrustManagerFactoryTest.class,
         X509KeyManagerTest.class,
 })
