+++ openjdk-integ-tests/src/test/java/org/conscrypt/javax/net/ssl/SSLEngineTest.java
@@ -17,27 +17,19 @@
 package org.conscrypt.javax.net.ssl;
 
 import static org.conscrypt.TestUtils.UTF_8;
-import static org.junit.Assert.assertArrayEquals;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertNotSame;
-import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.fail;
 
 import java.io.IOException;
 import java.nio.ByteBuffer;
 import java.util.Arrays;
-import java.util.concurrent.Callable;
-import java.util.concurrent.CountDownLatch;
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
-import java.util.concurrent.Future;
+import java.util.HashSet;
 import javax.crypto.SecretKey;
 import javax.crypto.spec.SecretKeySpec;
-import javax.net.ssl.HostnameVerifier;
-import javax.net.ssl.HttpsURLConnection;
 import javax.net.ssl.KeyManager;
 import javax.net.ssl.SSLContext;
 import javax.net.ssl.SSLEngine;
@@ -50,7 +42,6 @@ import javax.net.ssl.SSLParameters;
 import javax.net.ssl.SSLSession;
 import javax.net.ssl.X509ExtendedKeyManager;
 import libcore.java.security.StandardNames;
-import org.conscrypt.Conscrypt;
 import org.conscrypt.TestUtils;
 import org.conscrypt.java.security.TestKeyStore;
 import org.junit.Test;
@@ -213,11 +204,10 @@ public class SSLEngineTest {
                     continue;
                 }
                 /*
-                 * Kerberos cipher suites require external setup. See "Kerberos Requirements" in
-                 * https://java.sun.com/j2se/1.5.0/docs/guide/security/jsse/JSSERefGuide.html
-                 * #KRBRequire
+                 * This test uses TLS 1.2, and the TLS 1.3 cipher suites aren't customizable
+                 * anyway.
                  */
-                if (cipherSuite.startsWith("TLS_KRB5_")) {
+                if (StandardNames.CIPHER_SUITES_TLS13.contains(cipherSuite)) {
                     continue;
                 }
 
@@ -421,11 +411,38 @@ public class SSLEngineTest {
         e.setEnabledCipherSuites(e.getSupportedCipherSuites());
 
         // Check that setEnabledCipherSuites affects getEnabledCipherSuites
-        String[] cipherSuites = new String[] {e.getSupportedCipherSuites()[0]};
+        String[] cipherSuites = new String[] {
+                TestUtils.pickArbitraryNonTls13Suite(e.getSupportedCipherSuites())
+        };
         e.setEnabledCipherSuites(cipherSuites);
         assertEquals(Arrays.asList(cipherSuites), Arrays.asList(e.getEnabledCipherSuites()));
 
         c.close();
+    }
+
+    @Test
+    public void test_SSLEngine_setEnabledCipherSuites_TLS13() throws Exception {
+        SSLContext context = SSLContext.getInstance("TLSv1.3");
+        context.init(null, null, null);
+        SSLEngine e = context.createSSLEngine();
+        // The TLS 1.3 cipher suites should be enabled by default
+        assertTrue(new HashSet<String>(Arrays.asList(e.getEnabledCipherSuites()))
+                .containsAll(StandardNames.CIPHER_SUITES_TLS13));
+        // Disabling them should be ignored
+        e.setEnabledCipherSuites(new String[0]);
+        assertTrue(new HashSet<String>(Arrays.asList(e.getEnabledCipherSuites()))
+                .containsAll(StandardNames.CIPHER_SUITES_TLS13));
+
+        e.setEnabledCipherSuites(new String[] {
+                TestUtils.pickArbitraryNonTls13Suite(e.getSupportedCipherSuites())
+        });
+        assertTrue(new HashSet<String>(Arrays.asList(e.getEnabledCipherSuites()))
+                .containsAll(StandardNames.CIPHER_SUITES_TLS13));
+
+        // Disabling TLS 1.3 should disable the 1.3 cipher suites
+        e.setEnabledProtocols(new String[] { "TLSv1.2" });
+        assertFalse(new HashSet<String>(Arrays.asList(e.getEnabledCipherSuites()))
+                .containsAll(StandardNames.CIPHER_SUITES_TLS13));
     }
 
     @Test
@@ -512,55 +529,6 @@ public class SSLEngineTest {
         c.close();
     }
 
-    @Test
-    public void test_SSLEngine_beginHandshake() throws Exception {
-        TestSSLContext c = TestSSLContext.create();
-
-        try {
-            c.clientContext.createSSLEngine().beginHandshake();
-            fail();
-        } catch (IllegalStateException expected) {
-            // Ignored.
-        }
-        c.close();
-
-        TestSSLEnginePair p = TestSSLEnginePair.create(null);
-        assertConnected(p);
-        p.close();
-    }
-
-    @Test
-    public void test_SSLEngine_beginHandshake_noKeyStore() throws Exception {
-        TestSSLContext c = TestSSLContext.newBuilder()
-                                   .useDefaults(false)
-                                   .clientContext(SSLContext.getDefault())
-                                   .serverContext(SSLContext.getDefault())
-                                   .build();
-        SSLEngine[] p = null;
-        try {
-            // TODO Fix KnownFailure AlertException "NO SERVER CERTIFICATE FOUND"
-            // ServerHandshakeImpl.selectSuite should not select a suite without a required cert
-            p = TestSSLEnginePair.connect(c, null);
-            fail();
-        } catch (SSLHandshakeException expected) {
-            // Ignored.
-        } finally {
-            if (p != null) {
-                TestSSLEnginePair.close(p);
-            }
-        }
-        c.close();
-    }
-
-    @Test
-    public void test_SSLEngine_beginHandshake_noClientCertificate() throws Exception {
-        TestSSLContext c = TestSSLContext.create();
-        SSLEngine[] engines = TestSSLEnginePair.connect(c, null);
-        assertConnected(engines[0], engines[1]);
-        c.close();
-        TestSSLEnginePair.close(engines);
-    }
-
     // http://b/37078438
     @Test
     public void test_SSLEngine_beginHandshake_redundantCalls() throws Exception {
@@ -627,7 +595,7 @@ public class SSLEngineTest {
     @Test
     public void test_SSLEngine_setUseClientMode_afterHandshake() throws Exception {
         // can't set after handshake
-        TestSSLEnginePair pair = TestSSLEnginePair.create(null);
+        TestSSLEnginePair pair = TestSSLEnginePair.create();
         try {
             pair.server.setUseClientMode(false);
             fail();
@@ -659,122 +627,6 @@ public class SSLEngineTest {
                 server.setUseClientMode(serverClientMode);
             }
         }, finished);
-    }
-
-    @Test
-    public void test_SSLEngine_clientAuth() throws Exception {
-        TestSSLContext c = TestSSLContext.create();
-        SSLEngine e = c.clientContext.createSSLEngine();
-
-        assertFalse(e.getWantClientAuth());
-        assertFalse(e.getNeedClientAuth());
-
-        // confirm turning one on by itself
-        e.setWantClientAuth(true);
-        assertTrue(e.getWantClientAuth());
-        assertFalse(e.getNeedClientAuth());
-
-        // confirm turning setting on toggles the other
-        e.setNeedClientAuth(true);
-        assertFalse(e.getWantClientAuth());
-        assertTrue(e.getNeedClientAuth());
-
-        // confirm toggling back
-        e.setWantClientAuth(true);
-        assertTrue(e.getWantClientAuth());
-        assertFalse(e.getNeedClientAuth());
-
-        // TODO Fix KnownFailure "init - invalid private key"
-        TestSSLContext clientAuthContext = TestSSLContext.create(
-                TestKeyStore.getClientCertificate(), TestKeyStore.getServer());
-        TestSSLEnginePair p =
-                TestSSLEnginePair.create(clientAuthContext, new TestSSLEnginePair.Hooks() {
-                    @Override
-                    void beforeBeginHandshake(SSLEngine client, SSLEngine server) {
-                        server.setWantClientAuth(true);
-                    }
-                });
-        assertConnected(p);
-        assertNotNull(p.client.getSession().getLocalCertificates());
-        TestKeyStore.assertChainLength(p.client.getSession().getLocalCertificates());
-        TestSSLContext.assertClientCertificateChain(
-                clientAuthContext.clientTrustManager, p.client.getSession().getLocalCertificates());
-        clientAuthContext.close();
-        c.close();
-        p.close();
-    }
-
-    /**
-     * http://code.google.com/p/android/issues/detail?id=31903
-     * This test case directly tests the fix for the issue.
-     */
-    @Test
-    public void test_SSLEngine_clientAuthWantedNoClientCert() throws Exception {
-        TestSSLContext clientAuthContext =
-                TestSSLContext.create(TestKeyStore.getClient(), TestKeyStore.getServer());
-        TestSSLEnginePair p =
-                TestSSLEnginePair.create(clientAuthContext, new TestSSLEnginePair.Hooks() {
-                    @Override
-                    void beforeBeginHandshake(SSLEngine client, SSLEngine server) {
-                        server.setWantClientAuth(true);
-                    }
-                });
-        assertConnected(p);
-        clientAuthContext.close();
-        p.close();
-    }
-
-    /**
-     * http://code.google.com/p/android/issues/detail?id=31903
-     * This test case verifies that if the server requires a client cert
-     * (setNeedClientAuth) but the client does not provide one SSL connection
-     * establishment will fail
-     */
-    @Test
-    public void test_SSLEngine_clientAuthNeededNoClientCert() throws Exception {
-        TestSSLContext clientAuthContext =
-                TestSSLContext.create(TestKeyStore.getClient(), TestKeyStore.getServer());
-        TestSSLEnginePair p = null;
-        try {
-            p = TestSSLEnginePair.create(clientAuthContext, new TestSSLEnginePair.Hooks() {
-                @Override
-                void beforeBeginHandshake(SSLEngine client, SSLEngine server) {
-                    server.setNeedClientAuth(true);
-                }
-            });
-            fail();
-        } catch (SSLHandshakeException expected) {
-            // Ignored.
-        } finally {
-            clientAuthContext.close();
-            if (p != null) {
-                p.close();
-            }
-        }
-    }
-
-    @Test
-    public void test_SSLEngine_endpointVerification_Success() throws Exception {
-        TestUtils.assumeSetEndpointIdentificationAlgorithmAvailable();
-        // The default hostname verifier on OpenJDK just rejects all hostnames,
-        // which is not helpful, so replace with a basic functional one.
-        HostnameVerifier oldDefault = HttpsURLConnection.getDefaultHostnameVerifier();
-        HttpsURLConnection.setDefaultHostnameVerifier(new TestHostnameVerifier());
-        try {
-            TestSSLContext c = TestSSLContext.create();
-            TestSSLEnginePair p = TestSSLEnginePair.create(c, new TestSSLEnginePair.Hooks() {
-                @Override
-                void beforeBeginHandshake(SSLEngine client, SSLEngine server) {
-                    SSLParameters p = client.getSSLParameters();
-                    p.setEndpointIdentificationAlgorithm("HTTPS");
-                    client.setSSLParameters(p);
-                }
-            });
-            assertConnected(p);
-            c.close();
-        } finally {
-            HttpsURLConnection.setDefaultHostnameVerifier(oldDefault);
-        }
     }
 
     @Test
@@ -897,333 +749,6 @@ public class SSLEngineTest {
             assertFalse(e.getWantClientAuth());
         }
         c.close();
-    }
-
-    @Test
-    public void test_TestSSLEnginePair_create() throws Exception {
-        TestSSLEnginePair test = TestSSLEnginePair.create(null);
-        assertNotNull(test.c);
-        assertNotNull(test.server);
-        assertNotNull(test.client);
-        assertConnected(test);
-        test.close();
-    }
-
-    private final int NUM_STRESS_ITERATIONS = 1000;
-
-    @Test
-    public void test_SSLEngine_Multiple_Thread_Success() throws Exception {
-        final TestSSLEnginePair pair = TestSSLEnginePair.create();
-        try {
-            assertConnected(pair);
-
-            final CountDownLatch startUpSync = new CountDownLatch(2);
-            ExecutorService executor = Executors.newFixedThreadPool(2);
-            Future<Void> client = executor.submit(new Callable<Void>() {
-                @Override
-                public Void call() throws Exception {
-                    startUpSync.countDown();
-
-                    for (int i = 0; i < NUM_STRESS_ITERATIONS; i++) {
-                        assertSendsCorrectly("This is the client. Hello!".getBytes(UTF_8),
-                                pair.client, pair.server, false);
-                    }
-
-                    return null;
-                }
-            });
-            Future<Void> server = executor.submit(new Callable<Void>() {
-                @Override
-                public Void call() throws Exception {
-                    startUpSync.countDown();
-
-                    for (int i = 0; i < NUM_STRESS_ITERATIONS; i++) {
-                        assertSendsCorrectly("This is the server. Hi!".getBytes(UTF_8), pair.server,
-                                pair.client, false);
-                    }
-
-                    return null;
-                }
-            });
-            executor.shutdown();
-            client.get();
-            server.get();
-        } finally {
-            pair.close();
-        }
-    }
-
-    @Test
-    public void test_SSLEngine_CloseOutbound() throws Exception {
-        final TestSSLEnginePair pair = TestSSLEnginePair.create();
-        try {
-            assertConnected(pair);
-
-            // Closing the outbound direction should cause a close_notify to be sent
-            pair.client.closeOutbound();
-            ByteBuffer clientOut = ByteBuffer
-                    .allocate(pair.client.getSession().getPacketBufferSize());
-            SSLEngineResult res = pair.client.wrap(ByteBuffer.wrap(new byte[0]), clientOut);
-            assertEquals(Status.CLOSED, res.getStatus());
-            assertEquals(HandshakeStatus.NOT_HANDSHAKING, res.getHandshakeStatus());
-            assertTrue(res.bytesProduced() > 0);
-
-            // Read the close_notify in the server
-            clientOut.flip();
-            ByteBuffer serverIn = ByteBuffer
-                    .allocate(pair.server.getSession().getApplicationBufferSize());
-            res = pair.server.unwrap(clientOut, serverIn);
-            assertEquals(Status.CLOSED, res.getStatus());
-            assertEquals(HandshakeStatus.NEED_WRAP, res.getHandshakeStatus());
-
-            // Reading the close_notify should cause a close_notify to be sent back
-            ByteBuffer serverOut = ByteBuffer
-                    .allocate(pair.server.getSession().getPacketBufferSize());
-            res = pair.server.wrap(ByteBuffer.wrap(new byte[0]), serverOut);
-            assertEquals(Status.CLOSED, res.getStatus());
-            assertEquals(HandshakeStatus.NOT_HANDSHAKING, res.getHandshakeStatus());
-            assertTrue(res.bytesProduced() > 0);
-
-            // Read the close_notify in the client
-            serverOut.flip();
-            ByteBuffer clientIn = ByteBuffer
-                    .allocate(pair.client.getSession().getApplicationBufferSize());
-            res = pair.client.unwrap(serverOut, clientIn);
-            assertEquals(Status.CLOSED, res.getStatus());
-            assertEquals(HandshakeStatus.NOT_HANDSHAKING, res.getHandshakeStatus());
-
-            // Both sides have received close_notify messages, so both peers should have
-            // registered that they're finished
-            assertTrue(pair.client.isInboundDone() && pair.client.isOutboundDone());
-            assertTrue(pair.server.isInboundDone() && pair.server.isOutboundDone());
-        } finally {
-            pair.close();
-        }
-    }
-
-    @Test
-    public void test_SSLEngine_Closed() throws Exception {
-        TestSSLEnginePair pair = TestSSLEnginePair.create();
-        pair.close();
-        ByteBuffer out = ByteBuffer.allocate(pair.client.getSession().getPacketBufferSize());
-        SSLEngineResult res = pair.client.wrap(ByteBuffer.wrap(new byte[] { 0x01 }), out);
-        assertEquals(Status.CLOSED, res.getStatus());
-        assertEquals(0, res.bytesProduced());
-
-        res = pair.client.unwrap(ByteBuffer.wrap(new byte[] { 0x01} ), out);
-        assertEquals(Status.CLOSED, res.getStatus());
-        assertEquals(0, res.bytesConsumed());
-    }
-
-    @Test
-    public void test_SSLEngine_TlsUnique() throws Exception {
-        TestSSLEnginePair pair = TestSSLEnginePair.create(new TestSSLEnginePair.Hooks() {
-            @Override
-            void beforeBeginHandshake(SSLEngine client, SSLEngine server) {
-                assertNull(Conscrypt.getTlsUnique(client));
-                assertNull(Conscrypt.getTlsUnique(server));
-            }
-        });
-        try {
-            assertConnected(pair);
-
-            byte[] clientTlsUnique = Conscrypt.getTlsUnique(pair.client);
-            byte[] serverTlsUnique = Conscrypt.getTlsUnique(pair.server);
-            assertNotNull(clientTlsUnique);
-            assertNotNull(serverTlsUnique);
-            assertArrayEquals(clientTlsUnique, serverTlsUnique);
-        } finally {
-            pair.close();
-        }
-    }
-
-    @Test
-    public void test_SSLEngine_TokenBinding_Success() throws Exception {
-        TestSSLEnginePair pair = TestSSLEnginePair.create(new TestSSLEnginePair.Hooks() {
-            @Override
-            void beforeBeginHandshake(SSLEngine client, SSLEngine server) {
-                try {
-                    Conscrypt.setTokenBindingParams(client, 1, 2);
-                    Conscrypt.setTokenBindingParams(server, 2, 3);
-                } catch (SSLException e) {
-                    throw new RuntimeException(e);
-                }
-                assertEquals(-1, Conscrypt.getTokenBindingParams(client));
-                assertEquals(-1, Conscrypt.getTokenBindingParams(server));
-            }
-        });
-        try {
-            assertConnected(pair);
-
-            assertEquals(2, Conscrypt.getTokenBindingParams(pair.client));
-            assertEquals(2, Conscrypt.getTokenBindingParams(pair.server));
-        } finally {
-            pair.close();
-        }
-    }
-
-    @Test
-    public void test_SSLEngine_TokenBinding_NoClientSupport() throws Exception {
-        TestSSLEnginePair pair = TestSSLEnginePair.create(new TestSSLEnginePair.Hooks() {
-            @Override
-            void beforeBeginHandshake(SSLEngine client, SSLEngine server) {
-                try {
-                    // Do not enable on client
-                    Conscrypt.setTokenBindingParams(server, 2, 3);
-                } catch (SSLException e) {
-                    throw new RuntimeException(e);
-                }
-                assertEquals(-1, Conscrypt.getTokenBindingParams(client));
-                assertEquals(-1, Conscrypt.getTokenBindingParams(server));
-            }
-        });
-        try {
-            assertConnected(pair);
-
-            assertEquals(-1, Conscrypt.getTokenBindingParams(pair.client));
-            assertEquals(-1, Conscrypt.getTokenBindingParams(pair.server));
-        } finally {
-            pair.close();
-        }
-    }
-
-    @Test
-    public void test_SSLEngine_TokenBinding_NoServerSupport() throws Exception {
-        TestSSLEnginePair pair = TestSSLEnginePair.create(new TestSSLEnginePair.Hooks() {
-            @Override
-            void beforeBeginHandshake(SSLEngine client, SSLEngine server) {
-                try {
-                    // Do not enable on server
-                    Conscrypt.setTokenBindingParams(client, 2, 3);
-                } catch (SSLException e) {
-                    throw new RuntimeException(e);
-                }
-                assertEquals(-1, Conscrypt.getTokenBindingParams(client));
-                assertEquals(-1, Conscrypt.getTokenBindingParams(server));
-            }
-        });
-        try {
-            assertConnected(pair);
-
-            assertEquals(-1, Conscrypt.getTokenBindingParams(pair.client));
-            assertEquals(-1, Conscrypt.getTokenBindingParams(pair.server));
-        } finally {
-            pair.close();
-        }
-    }
-
-    @Test
-    public void test_SSLEngine_TokenBinding_MismatchedSupport() throws Exception {
-        TestSSLEnginePair pair = TestSSLEnginePair.create(new TestSSLEnginePair.Hooks() {
-            @Override
-            void beforeBeginHandshake(SSLEngine client, SSLEngine server) {
-                try {
-                    Conscrypt.setTokenBindingParams(client, 2);
-                    Conscrypt.setTokenBindingParams(server, 1, 3);
-                } catch (SSLException e) {
-                    throw new RuntimeException(e);
-                }
-                assertEquals(-1, Conscrypt.getTokenBindingParams(client));
-                assertEquals(-1, Conscrypt.getTokenBindingParams(server));
-            }
-        });
-        try {
-            assertConnected(pair);
-
-            assertEquals(-1, Conscrypt.getTokenBindingParams(pair.client));
-            assertEquals(-1, Conscrypt.getTokenBindingParams(pair.server));
-        } finally {
-            pair.close();
-        }
-    }
-
-    @Test
-    public void test_SSLEngine_TokenBinding_MismatchedOrdering() throws Exception {
-        // When the server and client disagree on the preference order, the server should
-        // select the server's most highly preferred value.
-        TestSSLEnginePair pair = TestSSLEnginePair.create(new TestSSLEnginePair.Hooks() {
-            @Override
-            void beforeBeginHandshake(SSLEngine client, SSLEngine server) {
-                try {
-                    Conscrypt.setTokenBindingParams(client, 1, 2, 3, 4);
-                    Conscrypt.setTokenBindingParams(server, 3, 2);
-                } catch (SSLException e) {
-                    throw new RuntimeException(e);
-                }
-                assertEquals(-1, Conscrypt.getTokenBindingParams(client));
-                assertEquals(-1, Conscrypt.getTokenBindingParams(server));
-            }
-        });
-        try {
-            assertConnected(pair);
-
-            assertEquals(3, Conscrypt.getTokenBindingParams(pair.client));
-            assertEquals(3, Conscrypt.getTokenBindingParams(pair.server));
-        } finally {
-            pair.close();
-        }
-    }
-
-    @Test
-    public void test_SSLEngine_TokenBinding_ExceptionAfterConnect() throws Exception {
-        TestSSLEnginePair pair = TestSSLEnginePair.create();
-        try {
-            assertConnected(pair);
-
-            try {
-                Conscrypt.setTokenBindingParams(pair.client, 1);
-                fail("setTokenBindingParams after handshake should throw");
-            } catch (IllegalStateException expected) {
-            }
-            try {
-                Conscrypt.setTokenBindingParams(pair.server, 1);
-                fail("setTokenBindingParams after handshake should throw");
-            } catch (IllegalStateException expected) {
-            }
-        } finally {
-            pair.close();
-        }
-    }
-
-    @Test
-    public void test_SSLEngine_EKM() throws Exception {
-        TestSSLEnginePair pair = TestSSLEnginePair.create(new TestSSLEnginePair.Hooks() {
-            @Override
-            void beforeBeginHandshake(SSLEngine client, SSLEngine server) {
-                try {
-                    assertNull(Conscrypt.exportKeyingMaterial(client, "FOO", null, 20));
-                    assertNull(Conscrypt.exportKeyingMaterial(server, "FOO", null, 20));
-                } catch (SSLException e) {
-                    throw new RuntimeException(e);
-                }
-            }
-        });
-        try {
-            assertConnected(pair);
-
-            byte[] clientEkm = Conscrypt.exportKeyingMaterial(pair.client, "FOO", null, 20);
-            byte[] serverEkm = Conscrypt.exportKeyingMaterial(pair.server, "FOO", null, 20);
-            assertNotNull(clientEkm);
-            assertNotNull(serverEkm);
-            assertEquals(20, clientEkm.length);
-            assertEquals(20, serverEkm.length);
-            assertArrayEquals(clientEkm, serverEkm);
-
-            byte[] clientContextEkm = Conscrypt.exportKeyingMaterial(
-                    pair.client, "FOO", new byte[0], 20);
-            byte[] serverContextEkm = Conscrypt.exportKeyingMaterial(
-                    pair.server, "FOO", new byte[0], 20);
-            assertNotNull(clientContextEkm);
-            assertNotNull(serverContextEkm);
-            assertEquals(20, clientContextEkm.length);
-            assertEquals(20, serverContextEkm.length);
-            assertArrayEquals(clientContextEkm, serverContextEkm);
-
-            // An empty context should be different than a null context
-            assertFalse(Arrays.equals(clientEkm, clientContextEkm));
-        } finally {
-            pair.close();
-        }
     }
 
     private void assertConnected(TestSSLEnginePair e) {
